extracts cache + rateLimit logic into utils to have proxyController focused on just proxy logic
