+++ server/utils/rateLimit.js
@@ -1,0 +1,8 @@
+const Bottleneck = require('bottleneck');
+const rateLimiter = new Bottleneck({
+  minTime: 1050 / 5,
+});
+
+module.exports = {
+  rateLimiter: rateLimiter,
+};
